requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
faiss-cpu
openai
python-dotenv
google-generativeai
pytest
httpx
